Extract out commonalities to form a superclass Contact.
